# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Medical Coding AI",
  "description": "AI-assisted medical coding: clinical documents -> evidence-backed codes -> validation -> human review.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/users": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_v1_users_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/UserOut"
         },
         "type": "array",
         "title": "Response List Users Api V1 Users Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ]
   },
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create User",
    "operationId": "create_user_api_v1_users_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ]
   }
  },
  "/api/v1/users/{user_id}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update User",
    "operationId": "update_user_api_v1_users__user_id__patch",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/api-keys": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Keys",
    "operationId": "list_keys_api_v1_api_keys_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ApiKeyOut"
         },
         "type": "array",
         "title": "Response List Keys Api V1 Api Keys Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ]
   },
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create Key",
    "operationId": "create_key_api_v1_api_keys_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ApiKeyCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ApiKeyCreated"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ]
   }
  },
  "/api/v1/api-keys/{key_id}": {
   "delete": {
    "tags": [
     "admin"
    ],
    "summary": "Revoke Key",
    "operationId": "revoke_key_api_v1_api_keys__key_id__delete",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "key_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Key Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/suggestions/{sid}": {
   "get": {
    "tags": [
     "review"
    ],
    "summary": "Get Suggestion",
    "operationId": "get_suggestion_api_v1_suggestions__sid__get",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "sid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Sid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuggestionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/suggestions/{sid}/approve": {
   "post": {
    "tags": [
     "review"
    ],
    "summary": "Approve",
    "operationId": "approve_api_v1_suggestions__sid__approve_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "sid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Sid"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/ApproveIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuggestionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/suggestions/{sid}/reject": {
   "post": {
    "tags": [
     "review"
    ],
    "summary": "Reject",
    "operationId": "reject_api_v1_suggestions__sid__reject_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "sid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Sid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RejectIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuggestionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/suggestions/{sid}/edit": {
   "post": {
    "tags": [
     "review"
    ],
    "summary": "Edit",
    "operationId": "edit_api_v1_suggestions__sid__edit_post",
    "security": [
     {
      "HTTPBearer": []
     },
     {
      "APIKeyHeader": []
     }
    ],
    "parameters": [
     {
      "name": "sid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Sid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EditIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuggestionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ApiKeyCreate": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 2,
      "title": "Name"
     },
     "role": {
      "type": "string",
      "enum": [
       "service",
       "auditor"
      ],
      "title": "Role",
      "default": "service"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "ApiKeyCreate"
   },
   "ApiKeyCreated": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "prefix": {
      "type": "string",
      "title": "Prefix"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "revoked": {
      "type": "boolean",
      "title": "Revoked"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_used_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Used At"
     },
     "key": {
      "type": "string",
      "title": "Key",
      "description": "Shown once. Store it securely."
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "prefix",
     "role",
     "revoked",
     "created_at",
     "key"
    ],
    "title": "ApiKeyCreated"
   },
   "ApiKeyOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "prefix": {
      "type": "string",
      "title": "Prefix"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "revoked": {
      "type": "boolean",
      "title": "Revoked"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_used_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Used At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "prefix",
     "role",
     "revoked",
     "created_at"
    ],
    "title": "ApiKeyOut"
   },
   "ApproveIn": {
    "properties": {
     "comment": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2000
       },
       {
        "type": "null"
       }
      ],
      "title": "Comment"
     }
    },
    "type": "object",
    "title": "ApproveIn"
   },
...
   "EditIn": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 10,
      "minLength": 3,
      "title": "Code"
     },
     "reason": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 3,
      "title": "Reason"
     },
     "error_category": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ErrorCategory"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "code",
     "reason"
    ],
    "title": "EditIn"
   },
...
   "ErrorCategory": {
    "type": "string",
    "enum": [
     "wrong_code",
     "missing_code",
     "extra_code",
     "wrong_specificity",
     "negation_error",
     "wrong_procedure",
     "wrong_terminology",
     "retrieval_failure",
     "reasoning_failure",
     "validation_failure",
     "other"
    ],
    "title": "ErrorCategory",
    "description": "Error-analysis taxonomy recorded on coder corrections."
   },
   "EvidenceOut": {
    "properties": {
     "quote": {
      "type": "string",
      "title": "Quote"
     },
     "start": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start"
     },
     "end": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "End"
     },
     "match_score": {
      "type": "number",
      "title": "Match Score"
     }
    },
    "type": "object",
    "required": [
     "quote",
     "start",
     "end",
     "match_score"
    ],
    "title": "EvidenceOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RejectIn": {
    "properties": {
     "reason": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 3,
      "title": "Reason"
     },
     "error_category": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ErrorCategory"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "reason"
    ],
    "title": "RejectIn"
   },
...
   "Role": {
    "type": "string",
    "enum": [
     "admin",
     "coder",
     "auditor",
     "service"
    ],
    "title": "Role"
   },
...
   "SuggestionOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "document_id": {
      "type": "string",
      "format": "uuid",
      "title": "Document Id"
     },
     "code_system": {
      "type": "string",
      "title": "Code System"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "kb_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Kb Version"
     },
     "entity_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Text"
     },
     "rationale": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rationale"
     },
     "sequence": {
      "type": "integer",
      "title": "Sequence"
     },
     "source": {
      "type": "string",
      "title": "Source"
     },
     "llm_confidence": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm Confidence"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "confidence_breakdown": {
      "additionalProperties": true,
      "type": "object",
      "title": "Confidence Breakdown"
     },
     "retrieval_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Retrieval Score"
     },
     "retrieval_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Retrieval Rank"
     },
     "validation_status": {
      "type": "string",
      "title": "Validation Status"
     },
     "validation_issues": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Validation Issues"
     },
     "review_route": {
      "type": "string",
      "title": "Review Route"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "final_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Final Code"
     },
     "final_description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Final Description"
     },
     "reviewed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reviewed At"
     },
     "evidence": {
      "items": {
       "$ref": "#/components/schemas/EvidenceOut"
      },
      "type": "array",
      "title": "Evidence"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "document_id",
     "code_system",
     "code",
     "description",
     "kb_version",
     "entity_text",
     "rationale",
     "sequence",
     "source",
     "llm_confidence",
     "confidence",
     "confidence_breakdown",
     "retrieval_score",
     "retrieval_rank",
     "validation_status",
     "validation_issues",
     "review_route",
     "status",
     "final_code",
     "final_description",
     "reviewed_at",
     "evidence",
     "created_at"
    ],
    "title": "SuggestionOut"
   },
...
   "UserCreate": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "maxLength": 200,
      "title": "Full Name",
      "default": ""
     },
     "password": {
      "type": "string",
      "maxLength": 72,
      "minLength": 12,
      "title": "Password"
     },
     "role": {
      "$ref": "#/components/schemas/Role",
      "default": "coder"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "UserCreate"
   },
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "role": {
      "$ref": "#/components/schemas/Role"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "email",
     "full_name",
     "role",
     "is_active",
     "created_at"
    ],
    "title": "UserOut"
   },
   "UserUpdate": {
    "properties": {
     "full_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "role": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Role"
       },
       {
        "type": "null"
       }
      ]
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     },
     "password": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 72,
        "minLength": 12
       },
       {
        "type": "null"
       }
      ],
      "title": "Password"
     }
    },
    "type": "object",
    "title": "UserUpdate"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   },
   "APIKeyHeader": {
    "type": "apiKey",
    "in": "header",
    "name": "X-API-Key"
   }
  }
 }
}